Inserts en las tablas de Roles, Camiones, Clientes, Tipo Cilindro,Zonas y Update de columna en tabla Cliente

+++ BD/dbo/Tables/Clientes.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[Clientes] (
     [ClientesID]       INT           IDENTITY (1, 1) NOT NULL,
     [ClientesCode]     VARCHAR (5)   NULL,
-    [Cedula]           INT           NULL,
+    [Cedula]           BIGINT        NULL,
     [RazonSocial]      VARCHAR (100) NULL,
     [Empresa]          VARCHAR (100) NULL,
     [ClientesRol]      VARCHAR (100) NULL,
